+++ package.json
@@ -10,6 +10,7 @@
     "@types/jest": "^27.0.1",
     "@types/node": "^16.7.13",
     "@types/react": "^18.0.0",
+    "@types/react-calendar": "^4.1.0",
     "@types/react-dom": "^18.0.0",
     "@types/react-icons": "^3.0.0",
     "@types/react-redux": "^7.1.25",
@@ -17,6 +18,7 @@
     "@types/shortid": "^0.0.29",
     "@types/styled-components": "^5.1.26",
     "react": "^18.2.0",
+    "react-calendar": "^4.6.0",
     "react-dom": "^18.2.0",
     "react-icons": "^4.10.1",
     "react-redux": "^8.1.1",
