--- src/components/common/Header.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Header = () => {
-  return <h1>To do List</h1>;
-};
-
-export default Header;
