+++ src/shared/Router.tsx
@@ -1,14 +1,12 @@
 import { BrowserRouter, Route, Routes } from "react-router-dom";
-import Home from "../pages/Home";
-import Detail from "../pages/Detail";
-import NonPage from "../pages/NonPage";
+import Home from "../pages/home/Home";
+import NonPage from "../pages//nonPage/NonPage";
 
 const Router = () => {
   return (
     <BrowserRouter>
       <Routes>
         <Route path="/" element={<Home />} />
-        {/* <Route path="detail/:id" element={<Detail />} /> */}
         <Route path="*" element={<NonPage />} />
       </Routes>
     </BrowserRouter>
